pyproject.toml
[project]
name = "trip-planner-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Saigelaw",email = "saigelo.peter@gmail.com"}
]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
django-cors-headers = "^4.3"
requests = "^2.32.5"
drf-spectacular = "^0.28.0"
polyline = "^2.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
pytest-django = "^4.8"
factory-boy = "^3.3"
faker = "^24.0"
pytest-mock = "^3.15.0"

[build-system]
requires = ["poetry-core^2.0.0"]
build-backend = "poetry.core.masonry.api"
